broken stuff for debug
